Freshly build all VLC contribs
The prebuilts for VLC are currently for the old 2.1.x release series and
thus far behind the ones that are used for 2.2.x and 3.0.x which we are
using for the nightlies.

Also VLC (probably) links to its own static contribs, so some of the
brew dependencies are invalid as they are freshly build inside VLC's
contrib system.